Add API call for Searching Books
Fetched API from openlibrary.org and added an event listener
on a search box at the top of the page. I also created an object
respensible for sorting the list alphabetically. Both things are
in the early stages and not fully complete. I also made some
adjustments to the style of the website.