Add more filtering Options
I added more filtering options and also corrected a big mistake
that was creating an error upon adding books to and filtering the
book list if it already contained more than 5 books. The issue was
how the iterations were implemented. I changed them to for...of loops
while ensuring only the books on the list were being iterated through.