Add additional information to book line items
I added an event listner to the dom root, and when you click on a line
item, if it's a book information line, it will expand and show
additional information. More information will be added soon. Also, I
am testing an iframe implementation, so I will push this update with the
iframe at the top of the page to view in the netlify app.