Improve overall look and functionality
Made layout changes and also general improvements to the way editing and
deleting books is handled to reduce errors. Before, I was getting
an error after editing or deleting a book, when I attempted to add
icons again. This was corrected by refreshing the book list with
addToPage(). This seemed like a temporary solution at first, but upon
testing, it was nice to have a visual representation to being done
with delting/editing.