Add color scheme side menu & add to local storage
Created a hover side menu for color scheme and placed it in the
options menu as apposed to having it's own section on the navbar. Also,
the color chosen now saves under userSettings in the local storage
object and then gets saved in the bookList object, so the user
will not have to constantly change the color scheme.