Add a favorite button and functionality
I added a star icon, that when clicked, changed the value of the
favorite property for that book to true. If the book does not have a
favorite property, it will add once with the value of false. I also
added a favorite button in the filters sections, that when presssed,
calls a method in the bookList object that filters the bookList by
favorites and adds those books to the page with addToPage(). Upon
toggling, the normal bookList will be displayed also using addToPage().