Fix Bugs for favorite feature
The favorites icon was being appended to two different locations and
placed in the same position in the parent div .bookInfo using
position absolute. This was creating issues when attempting to remove
the icon as the icon did not exist in one location but did exist in
another. Now, the icon is appended to the same location as a child of
.bookInfo. I also made improvments to the code for favorites. I also
created a function that creates an elements and gives it a class or id
name depending what is specified.