Add a preview screen for all and searched books
Created a function that displays all book elements in a preview
screen that that can be used to get additional information before
adding it to the booklist. From this screen, you can move forward
and backwards through the book list. This function takes in the
referenced book object, so it can be used for both searchedBooks and
bookList objects (or any other in the future);