Add tag functionality
Added a tag section under filters that gives more flexibility than
just a favorite button. There is a tag object in bookList.userSettings
that stores all tags and their colors. This is used for adding new
custom tags and pulling the tag name and adding it to the specified tag
key in selected book. It is also used to reference the color to apply
on the line item for the book in the book list. More general
improvements were done to the look of the page and functionality,
including some minor fixes for the spotlight section.