Fix Adding Custom Books and also fix scrollToBook()
Made a correction to the addToObject function to ensure userSettings
is only updated when it is provided as a parameter. This corrected
the error I was getting when adding custom books.

For scrollToBook(), instead of device width and height, I now reference
the elements width and height to ensure the elements are centered
correctly when using scrollToBook().