Create new tag system
The previous tag system only allowed user to add one tag and clicking
the tag would toggle between them. This system was not very useful
to those that need multiple parameters to categorize their books
properly.

Now, every book objects tag property is an array that holds arrays of
two values. One for the tag name and the other for it's color.
Clicking on the book list line item will show every tag, which if
clicked, will add that tag to the book objects tag property. There
is a visual representation for each books tags that have been added, on
the selection itself, and also on the bottom of each book in the book
list.