Improve usability of the spotlight section
I made a slight change to the spotlight section, so when a user clicks
a book in this section, it will scroll to the book in the book list,
you then get a visual queue to click it again, and this will show the
book in the preview box. This change was made for people that might use
the spotlight section to help navigate instead of previewing the book
by default.