Create a Demo Page and Track First Time Visits
I created a demo booklist that sits in demoBooklist.js. there is
now a function that, during initialization, checks if user is visiting
for the first time, if so, a visited: true key value is added to the
local storage and the demo book list is shown with a welcome message.

I also created a new file option that starts a brand new booklist, so
users can wipe the demo booklist, or any created booklist to start over.

A minor fix was also made to image covers, so http(s) is used in the
image link, as well as a replacement cover for books without an image.